backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
